remove useEffect from publiclayout
