+++ src/utils/functions/language-sync-wrapper/languageSyncWrapper.jsx
@@ -1,0 +1,37 @@
+import { useEffect } from 'react';
+import { useLocation, useNavigate, Outlet } from 'react-router-dom';
+import { useLocale } from '@/hooks/common/use-locale/useLocale';
+import { DEFAULT_LOCALE, LOCALES } from '@/const/common/locales';
+
+export const LanguageSyncWrapper = () => {
+    const { pathname, search } = useLocation();
+    const navigate = useNavigate();
+    const { currentLanguage, i18n } = useLocale();
+
+    useEffect(() => {
+        const segments = pathname.split('/').filter(Boolean);
+        const langInUrl = segments[0];
+        const isSupported = LOCALES.includes(langInUrl);
+        const currentLang = currentLanguage;
+
+        if (isSupported && langInUrl !== currentLang) {
+            i18n.changeLanguage(langInUrl);
+            return;
+        }
+
+        if (isSupported) return;
+
+        if (pathname === '/') {
+            if (currentLang !== DEFAULT_LOCALE) {
+                i18n.changeLanguage(DEFAULT_LOCALE);
+            }
+            return;
+        }
+
+        if (currentLang === DEFAULT_LOCALE) return;
+
+        navigate(`/${currentLang}${pathname}${search}`, { replace: true });
+    }, [pathname, search, navigate, i18n]);
+
+    return <Outlet />;
+};
